Time Pumping Calculation Errors
Changed the time around the main loop back to 0.5seconds

If the day changes in the middle of pumping, ignore that event and start
checking time as soon as the day information is saved.

Changed GetRTCCminute and GetRTCCsecond to mask 8 rather than 6 bits
since the report is in BCD rather than binary

Read both minute and second at the same command to avoid roll over
problems.

Changed time pumping calculation to deal better with cases where the
second, minute or hour rolls over to 0.

Reset warning now happens after 5 sec and there is an ! printed every
0.5sec until reset.